requirements.txt
anyio==4.11.0
bio==1.8.0
biopython==1.85
biothings_client==0.4.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
gprofiler-official==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
mygene==3.2.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
platformdirs==4.4.0
pooch==1.8.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
umap==0.1.1
urllib3==2.5.0
Werkzeug==3.1.3
